requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
pyannote.audio>=3.1.1,<3.5
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.11.0
huggingface_hub>=0.23.4
soundfile>=0.13.1
asteroid>=0.7.0
transformers>=4.57.0
numpy
speechbrain>=0.5.14
azure-storage-blob>=12.21.0
azure-cognitiveservices-speech>=1.43.0

